+++ src/MarketStat.Services/MarketStat.Services.Facts/MarketStat.Services.Facts.FactSalaryService/Validators/FactSalaryValidator.cs
@@ -22,7 +22,7 @@ public static class FactSalaryValidator
             throw new ArgumentException("BonusAmount cannot be negative.");
     }
     
-    public static void ValidateForUpdate(int salaryFactId, int dateId, int cityId, int employerId, int jobRoleId,
+    public static void ValidateForUpdate(long salaryFactId, int dateId, int cityId, int employerId, int jobRoleId,
         int employeeId, decimal salaryAmount, decimal bonusAmount)
     {
         if (salaryFactId <= 0)
